Agregado cambio en las retencions de los codigos 312 1.75% y la 344 a 2.75%

+++ workspace/recursos/src/main/resources/sql/create_impuesto.sql
@@ -44,7 +44,7 @@ create table SRI_RETENCION_RENTA
     RETENCION_ID BIGINT,
     NOMBRE varchar(256),
     CODIGO_SRI varchar(4),
-    PORCENTAJE integer,
+    PORCENTAJE decimal(13,4),
     DESCRIPCION varchar(512),
     FECHA_INICIO date,
     FECHA_FIN date,
